Lundi-s2

+++ libft/src/ft_strncpy.c
@@ -6,7 +6,7 @@
 /*   By: fde-monc <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/24 11:47:56 by fde-monc          #+#    #+#             */
-/*   Updated: 2015/11/25 19:50:37 by fde-monc         ###   ########.fr       */
+/*   Updated: 2015/11/29 15:22:58 by fde-monc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strncpy(char *dest, const char *src, size_t n)
 {
 	size_t	i;
-	int		src_l;
+	size_t	src_l;
 
 	src_l = ft_strlen(src);
 	i = 0;
